@@ -2,16 +2,15 @@ name: admin
 description: The Beeper landing page and console
 
 environment:
-  sdk: '>=2.10.0-0 <3.0.0'
+  sdk: '>=2.17.5 <3.0.0'
 
 dependencies:
   beeper_common:
     path: ../common
-  intl: ^0.16.1
-  rxdart: ^0.24.1
-  websocket: ^0.0.5
+  intl: ^0.17.0
+  rxdart: ^0.27.5
 
 dev_dependencies:
-  build_runner: ^1.10.0
-  build_web_compilers: ^2.11.0
-  pedantic: ^1.9.0
+  build_runner: ^2.2.0
+  build_web_compilers: ^3.2.4
+  pedantic: ^1.11.1
